fix: updated component structure
