+++ packages/roomkit-react/src/Prebuilt/components/Notifications/Notifications.tsx
@@ -1,23 +1,17 @@
 /* eslint-disable no-case-declarations */
-import React, { useCallback, useEffect } from 'react';
-import {
-  HMSNotificationTypes,
-  HMSRoleChangeRequest,
-  HMSRoomState,
-  selectIsLocalScreenShared,
-  selectRoomState,
-  useAwayNotifications,
-  useCustomEvent,
-  useHMSNotifications,
-  useHMSStore,
-} from '@100mslive/react-sdk';
-import { useRoomLayout } from '../../provider/roomLayoutProvider';
+import React, { useCallback } from 'react';
+import { HMSRoleChangeRequest, HMSRoomState, selectRoomState, useCustomEvent, useHMSStore } from '@100mslive/react-sdk';
 // @ts-ignore: No implicit Any
 import { ToastManager } from '../Toast/ToastManager';
+import { AutoplayBlockedModal } from './AutoplayBlockedModal';
 import { ChatNotifications } from './ChatNotifications';
-import { ErrorNotificationModal } from './ErrorNotificationModal';
+import { DeviceChangeNotifications } from './DeviceChangeNotifications';
+import { ErrorNotifications } from './ErrorNotifications';
 import { HandRaisedNotifications } from './HandRaisedNotifications';
+import { InitErrorModal } from './InitErrorModal';
+import { MessageNotifications } from './MessageNotifications';
 import { PeerNotifications } from './PeerNotifications';
+import { PermissionErrorNotificationModal } from './PermissionErrorModal';
 import { PollNotificationModal } from './PollNotificationModal';
 import { ReconnectNotifications } from './ReconnectNotifications';
 import { TrackBulkUnmuteModal } from './TrackBulkUnmuteModal';
@@ -25,25 +19,12 @@ import { TrackNotifications } from './TrackNotifications';
 import { TrackUnmuteModal } from './TrackUnmuteModal';
 import { TranscriptionNotifications } from './TranscriptionNotifications';
 // @ts-ignore: No implicit Any
-import { useIsNotificationDisabled, useSubscribedNotifications } from '../AppData/useUISettings';
-import { usePIPWindow } from '../PIP/usePIPWindow';
+import { useIsNotificationDisabled } from '../AppData/useUISettings';
 import { ROLE_CHANGE_DECLINED } from '../../common/constants';
 
-const notificationTypes = [
-  HMSNotificationTypes.CHANGE_TRACK_STATE_REQUEST,
-  HMSNotificationTypes.DEVICE_CHANGE_UPDATE,
-  HMSNotificationTypes.NEW_MESSAGE,
-];
-
 export function Notifications() {
-  const notification = useHMSNotifications(notificationTypes);
-  const subscribedNotifications = useSubscribedNotifications() || {};
   const roomState = useHMSStore(selectRoomState);
   const isNotificationDisabled = useIsNotificationDisabled();
-  const { showNotification } = useAwayNotifications();
-  const amIScreenSharing = useHMSStore(selectIsLocalScreenShared);
-  const logoURL = useRoomLayout()?.logo?.url;
-  const { pipWindow } = usePIPWindow();
 
   const handleRoleChangeDenied = useCallback((request: HMSRoleChangeRequest & { peerName: string }) => {
     ToastManager.addToast({
@@ -53,39 +34,6 @@ export function Notifications() {
   }, []);
 
   useCustomEvent({ type: ROLE_CHANGE_DECLINED, onEvent: handleRoleChangeDenied });
-
-  useEffect(() => {
-    if (!notification || isNotificationDisabled) {
-      return;
-    }
-    switch (notification.type) {
-      case HMSNotificationTypes.CHANGE_TRACK_STATE_REQUEST:
-        const track = notification.data?.track;
-        if (!notification.data.enabled) {
-          ToastManager.addToast({
-            title: `Your ${track.source} ${track.type} was muted by
-                ${notification.data.requestedBy?.name}.`,
-          });
-        }
-        break;
-      case HMSNotificationTypes.DEVICE_CHANGE_UPDATE:
-        ToastManager.addToast({
-          title: notification.message,
-        });
-        break;
-      case HMSNotificationTypes.NEW_MESSAGE:
-        if (amIScreenSharing && !notification.data?.ignored && !pipWindow) {
-          showNotification(`New message from ${notification.data.senderName}`, {
-            body: notification.data.message,
-            icon: logoURL,
-          });
-        }
-        break;
-      default:
-        break;
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [notification, subscribedNotifications.ERROR, subscribedNotifications.METADATA_UPDATED]);
 
   if (isNotificationDisabled) {
     return null;
@@ -98,8 +46,13 @@ export function Notifications() {
       <TrackNotifications />
       {roomState === HMSRoomState.Connected ? <PeerNotifications /> : null}
       <PollNotificationModal />
+      <MessageNotifications />
+      <DeviceChangeNotifications />
       <ReconnectNotifications />
-      <ErrorNotificationModal />
+      <ErrorNotifications />
+      <AutoplayBlockedModal />
+      <PermissionErrorNotificationModal />
+      <InitErrorModal />
       <ChatNotifications />
       <HandRaisedNotifications />
       <TranscriptionNotifications />
